chore: move azure logic to a separate file
